error template added, refactored error controller
