+++ pom.xml
@@ -51,8 +51,6 @@
 
     <build>
         <pluginManagement>
-
-
             <plugins>
                 <plugin>
                     <groupId>org.springframework.boot</groupId>
