+++ src/main/resources/templates/error/error.ftl
@@ -1,0 +1,15 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Error!</title>
+</head>
+<body>
+    <h2>Error Page</h2>
+    <div>Status code: <b> ${statusCode} </b> </div>
+    <div> Error Message: <b> ${errorMessage} </b> </div>
+</body>
+</html>
